Updated gitignores, minor cleanup, renamed toggle actions to "Toggle ..." from "Set ..."
