+++ obssuite/togglestreaming/pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.stream-pi</groupId>
-    <artifactId>obssuite_setstudiomodeaction</artifactId>
+    <artifactId>obssuite_togglestreamingaction</artifactId>
     <version>1.0.0</version>
 
     <build>
