+++ obssuite/togglestreaming/src/main/java/module-info.java
@@ -1,4 +1,4 @@
-module com.stream_pi.obssuite.setstudiomodeaction
+module com.stream_pi.obssuite.togglestreamingaction
 {
     requires com.stream_pi.action_api;
     requires com.stream_pi.util;
@@ -6,5 +6,5 @@ module com.stream_pi.obssuite.setstudiomodeaction
     requires obs.websocket.java;
     requires com.stream_pi.obssuite.motheraction;
 
-    provides com.stream_pi.action_api.externalplugin.ExternalPlugin with SetStudioMode.SetStudioMode;
+    provides com.stream_pi.action_api.externalplugin.ExternalPlugin with togglestreaming.ToggleStreaming;
 }
