+++ obssuite/togglestreaming/src/main/java/togglestreaming/ToggleStreaming.java
@@ -1,26 +1,19 @@
-package setstreaming;
-
-import java.util.ArrayList;
+package togglestreaming;
 
 import com.stream_pi.action_api.actionproperty.property.BooleanProperty;
-import com.stream_pi.action_api.actionproperty.property.Property;
-import com.stream_pi.action_api.actionproperty.property.Type;
-import com.stream_pi.action_api.externalplugin.NormalAction;
 import com.stream_pi.action_api.externalplugin.ToggleAction;
 import com.stream_pi.util.alert.StreamPiAlert;
 import com.stream_pi.util.alert.StreamPiAlertType;
 import com.stream_pi.util.exception.MinorException;
-import com.stream_pi.util.version.Version;
 
 import mother.motherconnection.MotherConnection;
-import net.twasi.obsremotejava.OBSRemoteController;
 
-public class SetStreaming extends ToggleAction
+public class ToggleStreaming extends ToggleAction
 {
 
-    public SetStreaming()
+    public ToggleStreaming()
     {
-        setName("Set Streaming");
+        setName("Toggle Streaming");
         setCategory("OBS");
         setVisibilityInServerSettingsPane(false);
         setAuthor("rnayabed");
