+++ obssuite/togglestudiomode/.gitignore
@@ -5,3 +5,4 @@ target/
 .factorypath
 .project
 .settings/
+*.iml
