+++ obssuite/togglestudiomode/src/main/java/togglestudiomode/ToggleStudioMode.java
@@ -1,4 +1,4 @@
-package SetStudioMode;
+package togglestudiomode;
 
 import com.stream_pi.action_api.actionproperty.property.Property;
 import com.stream_pi.action_api.actionproperty.property.Type;
@@ -6,16 +6,15 @@ import com.stream_pi.action_api.externalplugin.ToggleAction;
 import com.stream_pi.util.alert.StreamPiAlert;
 import com.stream_pi.util.alert.StreamPiAlertType;
 import com.stream_pi.util.exception.MinorException;
-import com.stream_pi.util.version.Version;
 
 import mother.motherconnection.MotherConnection;
 
-public class SetStudioMode extends ToggleAction
+public class ToggleStudioMode extends ToggleAction
 {
 
-    public SetStudioMode()
+    public ToggleStudioMode()
     {
-        setName("Set Studio Mode");
+        setName("Toggle Studio Mode");
         setCategory("OBS");
         setVisibilityInServerSettingsPane(false);
         setAuthor("rnayabed");
