+++ websiteaction/.gitignore
@@ -1,6 +1,6 @@
 .idea/
 target/
-WebsiteAction.iml
+*.iml
 .classpath
 .factorypath
 .project
